Handwashing Game
Today I got the hand-washing game functional. I got some placeholder art in for testing. The player can click and drag the arms back and forth, drag the soap over the hand to remove germs. The game ends after 20 seconds and gives you a rating of how you did. Right now it is just a percentage but this can be expanded into stars or something more visual.