Almost done!
I have gotten the runner game, the face touching game and the hand washing game functional. I am working on the hidden object game. It mostly works just have to add the win condition. The face touching game is a little wonky and I will work on polishing it if there is time